+++ inst/+arduinoio/lib/OctaveCoreLibrary.h
@@ -22,7 +22,7 @@ class OctaveCoreLibrary : public LibraryBase
   OctaveArduinoClass &occlass;
 public:
    OctaveCoreLibrary(OctaveArduinoClass &oc);
-   void commandHandler(uint8_t cmdID, uint8_t* inputs, uint8_t payload_size);
+   void commandHandler(uint8_t cmdID, uint8_t* inputs, uint16_t payload_size);
 };
