+++ inst/+arduinoio/lib/OctaveServoLibrary.cpp
@@ -44,7 +44,7 @@ OctaveServoLibrary::OctaveServoLibrary (OctaveArduinoClass &oc)
 }
 
 void
-OctaveServoLibrary::commandHandler (uint8_t cmdID, uint8_t* data, uint8_t datasz)
+OctaveServoLibrary::commandHandler (uint8_t cmdID, uint8_t* data, uint16_t datasz)
 {
   switch (cmdID)
     {
