+++ inst/+arduinoio/lib/OctaveShiftRegisterLibrary.cpp
@@ -296,7 +296,7 @@ OctaveShiftRegisterLibrary::OctaveShiftRegisterLibrary(OctaveArduinoClass &oc)
 }
 
 void
-OctaveShiftRegisterLibrary::commandHandler (uint8_t cmdID, uint8_t* data, uint8_t datasz)
+OctaveShiftRegisterLibrary::commandHandler (uint8_t cmdID, uint8_t* data, uint16_t datasz)
 {
   switch (cmdID)
     {
