+++ inst/+arduinoioaddons/+ExampleAddon/src/Echo.h
@@ -27,7 +27,7 @@ public:
     libName = "ExampleAddon/Echo";
     a.registerLibrary(this);
   }
-  void commandHandler(uint8_t cmdID, uint8_t* data, uint8_t datasz)
+  void commandHandler(uint8_t cmdID, uint8_t* data, uint16_t datasz)
   {
     switch (cmdID)
       {
