+++ inst/+arduinoioaddons/+PulseTrain/src/PulseTrainAddon.cpp
@@ -139,7 +139,7 @@ PulseTrainAddon::PulseTrainAddon(OctaveArduinoClass& a)
   a.registerLibrary(this);
 }
 
-void PulseTrainAddon::commandHandler(uint8_t cmdID, uint8_t* data, uint8_t datasz)
+void PulseTrainAddon::commandHandler(uint8_t cmdID, uint8_t* data, uint16_t datasz)
 {
   switch (cmdID)
     {
