+++ inst/+arduinoioaddons/+SimpleStepper/src/SimpleStepperAddon.cpp
@@ -370,7 +370,7 @@ SimpleStepperAddon::SimpleStepperAddon(OctaveArduinoClass& a)
   a.registerLibrary(this);
 }
 
-void SimpleStepperAddon::commandHandler(uint8_t cmdID, uint8_t* data, uint8_t datasz)
+void SimpleStepperAddon::commandHandler(uint8_t cmdID, uint8_t* data, uint16_t datasz)
 {
   switch (cmdID)
     {
